requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.6.1
joblib==1.4.2
matplotlib==3.8.2
seaborn==0.13.0
gunicorn==21.2.0
setuptools==69.0.3
python-dotenv==1.0.1
